Add files via upload
An unbeatable AI player plays a significant role in Tic-Tac-Toe game versus human player .
As part of my beginner internship with Codsoft, I implemented a Tic Tac Toe game in Python featuring an AI agent that is unbeatable by a human player. The AI uses the minimax algorithm to ensure that it either wins or forces a draw in every game. This project helped me gain a deeper understanding of AI strategies and improved my problem-solving skills. It was a great opportunity to apply core programming concepts, and I’m proud of successfully completing this task.

#codsoft